// cgocore-ui/src/app/features/models/reference/commodities/views/full/SpecialHandlingCodeFullView.ts
import { SpecialHandlingCodeForm } from 'src/app/features/models/reference/commodities/forms/SpecialHandlingCodeForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class SpecialHandlingCodeFullView {

id: number;
form: SpecialHandlingCodeForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}